UI clean

+++ Team2FinalProject/Assets/Materials/Shaders/StandardShaderWithGlowing.shader
@@ -165,7 +165,7 @@ Shader "Custom/StandardShaderWithGlowing"
 							float ndotl = clamp(dot(i.normal, l), 0, 1);
 							float LightNear = PointNearFar[count].x;
 							float LightFar = PointNearFar[count].y;
-							float4 LightColor = PointLightColor[count];
+							fixed4 LightColor = PointLightColor[count];
 							if (d > LightNear) 
 							{
 								if (d < LightFar)
